Fix three consecutive training crashes; add shared-memory L12 cache
Crash history and fixes applied:
1. Job 23733182 — OOM killed at batch 603: increase --mem 384G → 600G
2. Job 23749254 — empty log (Python stdout buffered in conda run subprocess):
   add PYTHONUNBUFFERED=1 to slurm/train.sh
3. Job 23750234 — TypeError: Got unsupported ScalarType BFloat16 in evaluate():
   cast mu/var .float() before .numpy()

Performance fixes (from tensor core utilisation analysis):
- Add torch.autocast("cuda", dtype=torch.bfloat16) to train_one_epoch and evaluate
- Remove explicit .float() casts in model.py that upcast fp16→fp32 before linears
- Re-enable L12 RAM cache (DISABLE_L12_CACHE removed); add static cache for
  DEM/LULC/soil/masks so __getitem__ does zero zarr reads for all static data

RAM fix for full-scale run (660 GB estimated without sharing):
- Add _preload_l12_to_shm(): rank 0 writes all L12 arrays to /dev/shm tmpfs
  as numpy memmaps before dataset construction; all 4 DDP ranks open the same
  files read-only so the OS shares one physical copy (~400 GB saved, peak ~258 GB)
- Add NCCL_TIMEOUT=3600 to guard against NCCL timeout during zarr stalls

+++ model.py
@@ -406,7 +406,7 @@ class SoilMoistureModel(nn.Module):
 
         pyramid    = torch.zeros(B, MAX_ACQ, 4, self.d_model, device=device)
 
-        flat_l12   = l12.reshape(B * MAX_ACQ, 196, self.d_model).float()
+        flat_l12   = l12.reshape(B * MAX_ACQ, 196, self.d_model)
         flat_valid = valid.reshape(B * MAX_ACQ)
 
         valid_idx  = flat_valid.nonzero(as_tuple=True)[0]
@@ -435,7 +435,7 @@ class SoilMoistureModel(nn.Module):
         token_valid: (B, 14, 14) bool | None — True = valid patch."""
         B  = l12.shape[0]
         tv = token_valid.reshape(B, 196) if token_valid is not None else None
-        return spatial_pyramid_pool(l12.float(), token_valid=tv, attn=attn)
+        return spatial_pyramid_pool(l12, token_valid=tv, attn=attn)
 
     def _get_target_spatial_tokens(self, batch: dict, B: int,
                                    device) -> tuple[torch.Tensor, torch.Tensor]:
@@ -448,7 +448,7 @@ class SoilMoistureModel(nn.Module):
             spatial_tokens : (B, 196, 768)
             use_s1         : (B,) bool
         """
-        spatial_tokens = batch["anchor_l12"].float().to(device)        # (B, 196, 768)
+        spatial_tokens = batch["anchor_l12"].to(device)                 # (B, 196, 768)
         anchor_orbit   = batch["anchor_orbit"].to(device)              # (B,) long: 0=S2,1=S1_asc,2=S1_desc
 
         # 2D spatial positional encoding
@@ -477,7 +477,7 @@ class SoilMoistureModel(nn.Module):
         Features are precomputed by precompute_terramind.py and loaded by the dataset.
         """
         def to_spatial(key: str) -> torch.Tensor:
-            return (batch[key].float().to(device)          # (B, 196, 768)
+            return (batch[key].to(device)                  # (B, 196, 768)
                     .reshape(B, 14, 14, self.d_model)
                     .permute(0, 3, 1, 2))                  # (B, 768, 14, 14)
